Implement healt checker into agent
This commit adds a health checker to the agent.

The heatthchecker is automatically started at startup.
The check interval is by default set to 5min but it can be modified by
setting the value of `health-check-interval` to another value. The unit
of time is second.

The health check is done by making requests to `/health` endpoint of the
server.

Every failure is logged into a file with a format:
[time of format RFC3339] console.redhat.com is unreachable

As soon as the console is reachable, only one log entry is added:
[time of format RFC3339] console.redhat.com is OK

The log file is recreated at each startup. The location is set to the
data folder of the agent.

RESOLVES: ECOPROJECT-2162

Signed-off-by: Cosmin Tupangiu <cosmin@redhat.com>