store: Add transactions
Transaction are needed because we have to leave the db in a clean state
if something happen while mutating data. Currently, we have only one
resource _source_ but, with the new multi source model, we'll need to
mutate two resources _agent_ and _source_ in the same operation.
This requires that the op is done in a transaction.

This commit adds transaction support into the store. The code is largly
based on the transaction implementation found in other RH service
(e.g. uhc-service-account) but is has been made simplier to fit our
needs.

The transactions are wrapped into a context and passed to store when
services are asking for _source_ repository (for now, we have only one.
Soon to be more). The store checks the context and if it finds a
transaction it will use it to create the source repository. Otherwise,
it uses its own db connection.

Methods for _commit_ and _rollback_ are added so a service can rollback
or commit a transaction if an error occurs.

Signed-off-by: Cosmin Tupangiu <cosmin@redhat.com>