agent/agent: Refactor run command
The run command has been refractor to make it simpler. The inventory
status method does not have a loop anymore and it is called from the
agent itself.
The health check state is checked before sending update requests.

The idea is to reorganize the "inventory" and "healthcheck" into a
stateless services and let the agent control everything.

This commit does not change the logic of the agent in any way, it just
tries to improve organization of the code and the readability.

Signed-off-by: Cosmin Tupangiu <cosmin@redhat.com>