agent/multi-source: Implementation of multi-source model in the agent
This commit is a breaking-change. It implementing the multi-source model
into agent.

The agent is nolonger bound to a specific sourceID. The sourceID is
computed from vCenter ID. The agents sends requests to two endpoints:
`/agents/status` and `/sources/status`. The agent endpoint is used to
send the status of the agent which is actually the status of the source
in the old model.
The `source` endpoint is used to send the inventory of the source only
when the inventory is up-to-date.
The flow is:
- agent starts up and is waiting for credentials. It sends a request to
  /agents endpoint with the new state.
- user enters the credentials and if agent obtained the inventory, it
  sends it to /source endpoint.

Signed-off-by: Cosmin Tupangiu <cosmin@redhat.com>