model: Switching from multi-source to single-source
This commit changes the model from multi-source to single-source (i.e
one source - one agent).
The underlying assumption is that a source **always has** an agent even in the on-prem
scenario.

Changes:
db: the relationship between source and agent is one2one.
api: agent api do not change only the main api changes to reflect the
logic of the new model.

Signed-off-by: Cosmin Tupangiu <cosmin@redhat.com>