auth/agent: add implementation for agent authentication with custom keys
This commit adds a new implementation for agent authentication. The
rationale of this implementation is that each organization has its own
signing key which can be removed if the account of the org is
terminated.
For now, we store the private keys in pg but the long term solution
would be to store them in Vault.
The key store is wrapped with a simple cache wrapper that return from
the cache the public keys as these keys are needed for each agent's
request.

The process:
- when the image is generated, the key is retreived from the store. If
  the key is not found (i.e. first time when generating an image for
  that organization) a new rsa private key is generated and saved.
- the key is used to signed a jwt for the agent.

Signed-off-by: Cosmin Tupangiu <cosmin@redhat.com>