Verify RHCOS sha256 sum after download
This commit adds a check of image's sha256.
The published sha256 sum must be provided along with the image url.

Along with this check, a bug was fixed when a failed download can leave
leftover into the file. For that the buffer is rewind.
If all the download failed, the file is removed from the disk to prevend
the planner restarts without trying againg to download the image.

Signed-off-by: Cosmin Tupangiu <cosmin@redhat.com>