Revert ISO initialization changes from commit 22b41549d5ec6c5316d7faf8fbc89e04a767cd19
Reverting the ISO initializer changes due to inability to create PVC
with ReadWriteMany access mode in the target environment.

Changes reverted:
- Remove initContainers with migration-planner-init
- Remove PersistentVolumeClaim for migration-planner-iso-storage
- Restore iso-storage volume to use emptyDir instead of PVC
- Restore ISO initialization in run command instead of init command
- Remove PVC-related template parameters

This restores the original behavior where ISO initialization happens
during the run command startup rather than in a separate init container.

Note: The ISO initializer implementation is kept as it will be needed
for future enhancements when proper storage solutions are available.

Signed-off-by: Cosmin Tupangiu <cosmin@redhat.com>