ECOPROJECT-3167 | ci: Fix the way to extract the git_tag
The Tekton pipeline files used an invalid template expression that wasn't being processed,
causing to receive a literal string as the image tag instead of the actual tag name.
Fix: Replaced the complex template expression with {{git_tag}} which
directly provides the tag name for tag push events in PipelinesAsCode.

Signed-off-by: Nir Argaman <nargaman@redhat.com>