ECOPROJECT-3343 | fix: Support upgrade from prod to current main
This PR modifies the goose migration scripts, so we can support the upgrade
from branch release_0.1.4-1. Which contained 0250926140958 but didn't contain
previous migration script so it wasn't possible to migrate to current DB.

This PR create a new migration scripts, which union the previuos scripts,
so upgrade and also new DB creation works.

Signed-off-by: Ondra Machacek <omachace@redhat.com>