ECOPROJECT-3667 | fix: user receives error when creating assessments
ECOPROJECT-3503 made assessments private for users removing the sharing
of the assessments at the organization level.
But, the PR missed to remove the DB constraint `"org_id_name" UNIQUE
CONSTRAINT, btree (org_id, name)` which introduced a bug when a user
receives an error when creating an assessment even if he does not own an
assessment with that name.

This PR removes that constraint and add a constraint between
(username,assessment_name).

Signed-off-by: Cosmin Tupangiu <cosmin@redhat.com>